requirements.txt
langchain
openai
python-dotenv
pypdf
streamlit